Issue #220:updated deleteBooking

+++ service/src/main/java/com/myhome/services/springdatajpa/BookingSDJpaService.java
@@ -1,7 +1,12 @@
 package com.myhome.services.springdatajpa;
 
+import com.myhome.domain.Amenity;
+import com.myhome.domain.AmenityBookingItem;
 import com.myhome.repositories.AmenityBookingItemRepository;
+import com.myhome.repositories.AmenityRepository;
 import com.myhome.services.BookingService;
+import java.util.Optional;
+import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,13 +16,20 @@ public class BookingSDJpaService implements BookingService {
 
   private final AmenityBookingItemRepository bookingRepository;
 
+  @Transactional
   @Override
-  public boolean deleteBooking(String bookingId) {
-    return bookingRepository.findByAmenityBookingItemId(bookingId)
-        .map(bookingItem -> {
-          bookingRepository.delete(bookingItem);
-          return true;
-        })
-        .orElse(false);
+  public boolean deleteBooking(String amenityId, String bookingId) {
+    Optional<AmenityBookingItem> booking =
+        bookingRepository.findByAmenityBookingItemId(bookingId);
+    return booking.map(bookingItem -> {
+      boolean amenityFound =
+          bookingItem.getAmenity().getAmenityId().equals(amenityId);
+      if (amenityFound) {
+        bookingRepository.delete(bookingItem);
+        return true;
+      } else {
+        return false;
+      }
+    }).orElse(false);
   }
 }
